tech(singlestat): move singlestat test to plugin and change to typescript
